Moved core functionality to the core folder
