+++ signal-sweep/src/signal_sweep/core/signal_stream.py
