+++ signal-sweep/src/signal_sweep/shared/constants.py
@@ -1,4 +1,4 @@
-from ..handlers.text_handler import TextHandler
-from .source import SourceType
+from ..core.handlers.text_handler import TextHandler
+from ..core.source import SourceType
 
 SOURCE_TYPE_TO_HANDLER_DICT = {SourceType.TXT: TextHandler}
